API: Move vim_to_object to helpers.c
